# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Daton Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/{orgId}/units:
    get:
      operationId: listUnits
      tags: [units]
      summary: List units for an organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      operationId: createUnit
      tags: [units]
      summary: Create a unit
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUnitBody"
      responses:
        "201":
          description: Unit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /organizations/{orgId}/units/{unitId}:
    get:
      operationId: getUnit
      tags: [units]
      summary: Get unit details
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: unitId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unit details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    patch:
      operationId: updateUnit
      tags: [units]
      summary: Update a unit
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: unitId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUnitBody"
      responses:
        "200":
          description: Updated unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    delete:
      operationId: deleteUnit
      tags: [units]
      summary: Delete a unit
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: unitId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Unit:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        name:
          type: string
        code:
          type: ["string", "null"]
        type:
          type: string
          enum: [sede, filial]
        cnpj:
          type: ["string", "null"]
        status:
          type: string
          enum: [ativa, inativa]
        cep:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        streetNumber:
          type: ["string", "null"]
        neighborhood:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - organizationId
        - name
        - type
        - status
        - createdAt
        - updatedAt
    CreateUnitBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        type:
          type: string
          enum: [sede, filial]
        cnpj:
          type: string
        status:
          type: string
          enum: [ativa, inativa]
        cep:
          type: string
        address:
          type: string
        streetNumber:
          type: string
        neighborhood:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        phone:
          type: string
      required:
        - name
        - type
    UpdateUnitBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        type:
          type: string
          enum: [sede, filial]
        cnpj:
          type: string
        status:
          type: string
          enum: [ativa, inativa]
        cep:
          type: string
        address:
          type: string
        streetNumber:
          type: string
        neighborhood:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        phone:
          type: string
...